Revert "Issue 7669 7730 7739 7740 nav tool"
